requirements.txt
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
face-recognition==1.3.0
numpy==1.22.0
gtts==2.2.3
playsound==1.3.0
openai==0.27.0
SpeechRecognition==3.8.1
eel==0.14.0
google-generativeai